fix(content): authenticate finalization targets
